Banish List::Util
Oh, I know it's a core Perl module.  The problem is that sometimes it
doesn't work.  It's not clear whether this is because the people were
doing something that wasn't supposed to work, but if so, it should have
detected that.  And the CPAN RT listing doesn't look very promising.

Perl, you have failed me.  Maybe time to open another ticket.  Maybe.